Big update (check changes to better understand).
