+++ config.mk
@@ -1,0 +1,9 @@
+export OBJ_DIR=$(abspath obj)
+export BUILD_DIR=$(abspath build)
+export ETHCORE_SRC_DIR=$(abspath ethcore)
+export OUT_DIR=$(abspath out)
+export GRUB_CFG_DIR=$(abspath grub)
+
+export ASM=nasm
+export CC=cc
+export LD=ld
