+++ ethcore/boot/gdt.s
@@ -1,4 +1,5 @@
 global gdt_flush
+global tss_flush
 
 gdt_flush:
     MOV eax, [esp+4]
