+++ ethcore/ethcore.mk
@@ -1,0 +1,3 @@
+export ASMFLAGS=-f elf32
+export CFLAGS=-m32 -fno-stack-protector -fno-builtin -nostdlib
+export LDFLAGS=-m elf_i386
