+++ ethcore/kernel/timer.h
@@ -1,0 +1,9 @@
+#ifndef __K_TIMER_H__
+#define __K_TIMER_H__
+
+struct InterruptRegisters;
+
+void initTimer();
+void onIrq0(struct InterruptRegisters *regs);
+
+#endif // __K_TIMER_H__
