+++ ethcore/klib/kbool.h
@@ -1,0 +1,9 @@
+#ifndef __K_KLIB_BOOL_H__
+#define __K_KLIB_BOOL_H__
+
+#define bool unsigned char
+
+#define true 1
+#define false 0
+
+#endif // __K_KLIB_BOOL_H__
