+++ ethcore/klib/kio.h
@@ -1,0 +1,8 @@
+#ifndef __K_KLIB_IO_H__
+#define __K_KLIB_IO_H__
+
+void __kputc(char c);
+void __kputs(const char *str);
+void __kprintf(const char *fmt, ...);
+
+#endif // __K_KLIB_IO_H__
