+++ ethcore/klib/kwchar.h
@@ -1,0 +1,11 @@
+#ifndef __K_KLIB_WCHAR_H__
+#define __K_KLIB_WCHAR_H__
+
+#include "klib/kint.h"
+
+/* NOTE: I'm not sure if this must be `int32_t` or `uint32_t`.
+ * Fix me if I'm wrong.
+ */
+typedef uint32_t wchar_t;
+
+#endif // __K_KLIB_WCHAR_H__
