+++ ethcore/linker.ld
