--- ethos/boot/grub/grub.cfg
@@ -1,4 +1,0 @@
-menuentry "ethos" {
-    multiboot /boot/kernel/kernel
-    boot
-}
