+++ grub/grub.cfg
@@ -1,0 +1,4 @@
+menuentry "ethos" {
+  multiboot /boot/ethcore
+  boot
+}
