--- src/keyboard.h
@@ -1,3 +1,0 @@
-void initKeyboard();
-void keyboardHandler(struct InterruptRegisters *regs);
-void append(char *part);
