--- src/timer.h
@@ -1,2 +1,0 @@
-void initTimer();
-void onIrq0(struct InterruptRegisters *regs);
